Handle Empty Spring-Boot-Lib Directory
Previously, as part of enumerating all of the dependencies in a Spring Boot
application, the code made an assumption that there would always be something
in the directory identified by Spring-Boot-Lib.  While this is likely to be
true in practice, edge case testing by the Spring Boot team has found that the
buildpack would fail if there was not.  This change updates the buildpack to
handle an empty Spring-Boot-Lib directory gracefully.

[resolves #5]

Signed-off-by: Ben Hale <bhale@pivotal.io>